Fix exception when message definition contains duplicate types (#45)
Fix exception when message definition contains duplicate types

This PR removes duplicate types (with the same name) to ensure that at max one result is returned when searching a type by its name. This fixes an exception that occurred when the message definition contained duplicate types due to a bug in mcap convert.