ROS2: Normalize `byte` as `uint8` instead of `int8` (#49)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

### Description

In
[ROS2](https://docs.ros.org/en/iron/Concepts/Basic/About-Interfaces.html#field-types),
`byte` fields should be treated as `uint8` types:


> | Type name | C++     | Python          | DDS type |
> |-----------|---------|-----------------|----------|
> | byte      | uint8_t | builtins.bytes* | octet    |

Prior to this change we were treating `byte` fields as `int8` which
causes errors when parsing a byte field with a default value (e.g.
`255`) which is out of the range of a int8.